+++ src/main/java/chess/view/OutputView.java
@@ -37,6 +37,10 @@ public class OutputView {
         System.out.println("잘못된 커맨드입니다.");
     }
 
+    public static void printWrongPositionMessage() {
+        System.out.println("이동할 수 없습니다.");
+    }
+
     public static void printScore(ChessBoard chessBoard) {
         Team nowPlayingTeam = chessBoard.getNowPlayingTeam();
         String nowPlayingTeamName = nowPlayingTeam.getTeamName();
